+++ examples/default/models_auto.py
@@ -34,8 +34,8 @@ class TEmployee(typing.Protocol):
     query: orm.Query
 
     # Model properties
-    id: typing.Optional[int]
-    name: str
+    id: "sqlalchemy.Column[typing.Optional[int]]"
+    name: "sqlalchemy.Column[str]"
 
     def __init__(self, id: typing.Optional[int] = None, name: str = "Unknown") -> None:
         """
@@ -97,4 +97,4 @@ class TEmployee(typing.Protocol):
         ...
 
 
-Employee: TEmployee = models.Employee  # type: ignore
+Employee: typing.Type[TEmployee] = models.Employee  # type: ignore