+++ examples/relationship/one_to_many/models_auto.py
@@ -36,9 +36,9 @@ class TDivision(typing.Protocol):
     query: orm.Query
 
     # Model properties
-    id: typing.Optional[int]
-    name: typing.Optional[str]
-    employees: typing.Sequence["TEmployee"]
+    id: "sqlalchemy.Column[typing.Optional[int]]"
+    name: "sqlalchemy.Column[typing.Optional[str]]"
+    employees: 'sqlalchemy.Column[typing.Sequence["TEmployee"]]'
 
     def __init__(
         self,
@@ -110,7 +110,7 @@ class TDivision(typing.Protocol):
         ...
 
 
-Division: TDivision = models.Division  # type: ignore
+Division: typing.Type[TDivision] = models.Division  # type: ignore
 
 
 class EmployeeDict(typing.TypedDict, total=False):
@@ -138,8 +138,8 @@ class TEmployee(typing.Protocol):
     query: orm.Query
 
     # Model properties
-    id: typing.Optional[int]
-    name: typing.Optional[str]
+    id: "sqlalchemy.Column[typing.Optional[int]]"
+    name: "sqlalchemy.Column[typing.Optional[str]]"
 
     def __init__(
         self, id: typing.Optional[int] = None, name: typing.Optional[str] = None
@@ -203,4 +203,4 @@ class TEmployee(typing.Protocol):
         ...
 
 
-Employee: TEmployee = models.Employee  # type: ignore
+Employee: typing.Type[TEmployee] = models.Employee  # type: ignore