+++ open_alchemy/column_factory/__init__.py
@@ -6,33 +6,34 @@ import typing
 from open_alchemy import exceptions
 from open_alchemy import facades
 from open_alchemy import helpers
-from open_alchemy import types
+from open_alchemy import types as oa_types
 
 from . import array_ref
 from . import column
 from . import object_ref
 from . import read_only
+from . import types
 
 _REF_PATTER = re.compile(r"^#\/components\/schemas\/(\w+)$")
 
 
+_TReturnSchema = typing.Union[
+    oa_types.ColumnSchema,
+    oa_types.ObjectRefSchema,
+    oa_types.ArrayRefSchema,
+    oa_types.ReadOnlySchema,
+]
+
+
 def column_factory(
     *,
-    schema: types.Schema,
-    schemas: types.Schemas,
+    schema: oa_types.Schema,
+    schemas: oa_types.Schemas,
     required: typing.Optional[bool] = None,
     logical_name: str,
     model_name: str,
-    model_schema: types.Schema,
-) -> typing.Tuple[
-    typing.List[typing.Tuple[str, facades.sqlalchemy.column.Column]],
-    typing.Union[
-        types.ColumnSchema,
-        types.ObjectRefSchema,
-        types.ArrayRefSchema,
-        types.ReadOnlySchema,
-    ],
-]:
+    model_schema: oa_types.Schema,
+) -> typing.Tuple[types.TReturnValue, _TReturnSchema]:
     """
     Generate column based on OpenAPI schema property.
 
@@ -53,7 +54,7 @@ def column_factory(
     type_ = helpers.peek.type_(schema=schema, schemas=schemas)
 
     if type_ not in {"object", "array"}:
-        column_schema, column_value = column.handle_column(
+        column_value, column_schema = column.handle_column(
             schema=schema, schemas=schemas, required=required
         )
         return ([(logical_name, column_value)], column_schema)