+++ open_alchemy/table_args/factory.py
@@ -21,7 +21,9 @@ _COMMON_SCHEMAS_FILE = os.path.join(_DIRECTORY, *_PATHS, "common-schemas.json")
 
 
 def _spec_to_schema_name(
-    *, spec: types.AnyUnique, schema_names: typing.Optional[typing.List[str]] = None
+    *,
+    spec: typing.Union[types.AnyUnique, types.AnyIndex],
+    schema_names: typing.Optional[typing.List[str]] = None
 ) -> str:
     """
     Convert a specification to the name of the matched schema.
@@ -137,7 +139,10 @@ def _construct_unique(spec: types.Unique) -> schema.UniqueConstraint:
         The unique constraints.
 
     """
-    return schema.UniqueConstraint(*spec["columns"], name=spec.get("name"))
+    name = spec.get("name")
+    columns = spec["columns"]
+
+    return schema.UniqueConstraint(*columns, name=name)
 
 
 def _construct_index(spec: types.Index) -> schema.Index:
@@ -151,9 +156,14 @@ def _construct_index(spec: types.Index) -> schema.Index:
         The composite index.
 
     """
-    return schema.Index(
-        spec.get("name"), *spec["expressions"], unique=spec.get("unique")
-    )
+    # There is a bug in the sqlalchemy-stubs where name is not Optional for Index
+    name: str = spec.get("name")  # type: ignore
+    unique = spec.get("unique")
+    expressions = spec["expressions"]
+
+    if unique is not None:
+        return schema.Index(name, *expressions, unique=unique)
+    return schema.Index(name, *expressions)
 
 
 def unique_factory(
@@ -173,7 +183,7 @@ def unique_factory(
     return map(_construct_unique, mapped_spec)
 
 
-def index_factory(*, spec: types.AnyUnique) -> typing.Iterator[schema.Index]:
+def index_factory(*, spec: types.AnyIndex) -> typing.Iterator[schema.Index]:
     """
     Generate composite indexes from specification.