+++ open_alchemy/types.py
@@ -35,11 +35,38 @@ class GetBase(Protocol):
 # Unique consraint types
 ColumnList = typing.List[str]
 ColumnListList = typing.List[ColumnList]
-Unique = typing.Dict[str, typing.Any]
+
+
+class _UniqueBase(TypedDict, total=True):
+    """Base class for unique schema."""
+
+    columns: typing.List[str]
+
+
+class Unique(_UniqueBase, total=False):
+    """Unique schema."""
+
+    name: typing.Optional[str]
+
+
 UniqueList = typing.List[Unique]
 AnyUnique = typing.Union[ColumnList, ColumnListList, Unique, UniqueList]
 # Index types
-Index = typing.Dict[str, typing.Any]
+
+
+class _IndexBase(TypedDict, total=True):
+    """Base class for index schema."""
+
+    expressions: typing.List[str]
+
+
+class Index(_IndexBase, total=False):
+    """Index schema."""
+
+    name: typing.Optional[str]
+    unique: bool
+
+
 IndexList = typing.List[Index]
 AnyIndex = typing.Union[ColumnList, ColumnListList, Index, IndexList]
 # Type for the default value