+++ tests/open_alchemy/models_file/test_integration.py
@@ -3,6 +3,7 @@
 import sys
 
 import pytest
+from mypy import api
 
 from open_alchemy import models_file
 
@@ -57,7 +58,7 @@ class TModel({_EXPECTED_MODEL_BASE}):
     query: orm.Query
 
     # Model properties
-    id: typing.Optional[int]
+    id: "sqlalchemy.Column[typing.Optional[int]]"
 
     def __init__(self, id: typing.Optional[int] = None) -> None:
         """
@@ -115,7 +116,111 @@ class TModel({_EXPECTED_MODEL_BASE}):
         ...
 
 
-Model: TModel = models.Model  # type: ignore
+Model: typing.Type[TModel] = models.Model  # type: ignore
+''',
+        ),
+        (
+            [
+                (
+                    {
+                        "properties": {
+                            "ref_model": {"type": "object", "x-de-$ref": "RefModel"}
+                        }
+                    },
+                    "Model",
+                )
+            ],
+            f'''{_DOCSTRING}
+# pylint: disable=no-member,super-init-not-called,unused-argument
+
+import typing
+
+import sqlalchemy{_ADDITIONAL_IMPORT}
+from sqlalchemy import orm
+
+from open_alchemy import models
+
+
+class ModelDict({_EXPECTED_TD_BASE}, total=False):
+    """TypedDict for properties that are not required."""
+
+    ref_model: typing.Optional["RefModelDict"]
+
+
+class TModel({_EXPECTED_MODEL_BASE}):
+    """
+    SQLAlchemy model protocol.
+
+    Attrs:
+        ref_model: The ref_model of the Model.
+
+    """
+
+    # SQLAlchemy properties
+    __table__: sqlalchemy.Table
+    __tablename__: str
+    query: orm.Query
+
+    # Model properties
+    ref_model: 'sqlalchemy.Column[typing.Optional["TRefModel"]]'
+
+    def __init__(self, ref_model: typing.Optional["TRefModel"] = None) -> None:
+        """
+        Construct.
+
+        Args:
+            ref_model: The ref_model of the Model.
+
+        """
+        ...
+
+    @classmethod
+    def from_dict(cls, ref_model: typing.Optional["RefModelDict"] = None) -> "TModel":
+        """
+        Construct from a dictionary (eg. a POST payload).
+
+        Args:
+            ref_model: The ref_model of the Model.
+
+        Returns:
+            Model instance based on the dictionary.
+
+        """
+        ...
+
+    @classmethod
+    def from_str(cls, value: str) -> "TModel":
+        """
+        Construct from a JSON string (eg. a POST payload).
+
+        Returns:
+            Model instance based on the JSON string.
+
+        """
+        ...
+
+    def to_dict(self) -> ModelDict:
+        """
+        Convert to a dictionary (eg. to send back for a GET request).
+
+        Returns:
+            Dictionary based on the model instance.
+
+        """
+        ...
+
+    def to_str(self) -> str:
+        """
+        Convert to a JSON string (eg. to send back for a GET request).
+
+        Returns:
+            JSON string based on the model instance.
+
+        """
+        ...
+
+
+Model: typing.Type[TModel] = models.Model  # type: ignore
 ''',
         ),
         (
@@ -155,7 +260,7 @@ class TModel1({_EXPECTED_MODEL_BASE}):
     query: orm.Query
 
     # Model properties
-    id: typing.Optional[int]
+    id: "sqlalchemy.Column[typing.Optional[int]]"
 
     def __init__(self, id: typing.Optional[int] = None) -> None:
         """
@@ -213,7 +318,7 @@ class TModel1({_EXPECTED_MODEL_BASE}):
         ...
 
 
-Model1: TModel1 = models.Model1  # type: ignore
+Model1: typing.Type[TModel1] = models.Model1  # type: ignore
 
 
 class Model2Dict({_EXPECTED_TD_BASE}, total=False):
@@ -237,7 +342,7 @@ class TModel2({_EXPECTED_MODEL_BASE}):
     query: orm.Query
 
     # Model properties
-    id: typing.Optional[str]
+    id: "sqlalchemy.Column[typing.Optional[str]]"
 
     def __init__(self, id: typing.Optional[str] = None) -> None:
         """
@@ -295,7 +400,7 @@ class TModel2({_EXPECTED_MODEL_BASE}):
         ...
 
 
-Model2: TModel2 = models.Model2  # type: ignore
+Model2: typing.Type[TModel2] = models.Model2  # type: ignore
 ''',
         ),
         (
@@ -314,7 +419,7 @@ from open_alchemy import models
 class ModelDict({_EXPECTED_TD_BASE}, total=False):
     """TypedDict for properties that are not required."""
 
-    extremely_long_name_that_will_cause_wrapping_aaaaaaaaaaaaaaaaa: typing.Optional[int]
+    {_LONG_NAME}: typing.Optional[int]
 
 
 class TModel({_EXPECTED_MODEL_BASE}):
@@ -322,8 +427,8 @@ class TModel({_EXPECTED_MODEL_BASE}):
     SQLAlchemy model protocol.
 
     Attrs:
-        extremely_long_name_that_will_cause_wrapping_aaaaaaaaaaaaaaaaa: The
-            extremely_long_name_that_will_cause_wrapping_aaaaaaaaaaaaaaaaa of
+        {_LONG_NAME}: The
+            {_LONG_NAME} of
             the Model.
 
     """
@@ -334,11 +439,11 @@ class TModel({_EXPECTED_MODEL_BASE}):
     query: orm.Query
 
     # Model properties
-    extremely_long_name_that_will_cause_wrapping_aaaaaaaaaaaaaaaaa: typing.Optional[int]
+    {_LONG_NAME}: "sqlalchemy.Column[typing.Optional[int]]"
 
     def __init__(
         self,
-        extremely_long_name_that_will_cause_wrapping_aaaaaaaaaaaaaaaaa: typing.Optional[
+        {_LONG_NAME}: typing.Optional[
             int
         ] = None,
     ) -> None:
@@ -346,8 +451,8 @@ class TModel({_EXPECTED_MODEL_BASE}):
         Construct.
 
         Args:
-            extremely_long_name_that_will_cause_wrapping_aaaaaaaaaaaaaaaaa: The
-                extremely_long_name_that_will_cause_wrapping_aaaaaaaaaaaaaaaaa
+            {_LONG_NAME}: The
+                {_LONG_NAME}
                 of the Model.
 
         """
@@ -356,7 +461,7 @@ class TModel({_EXPECTED_MODEL_BASE}):
     @classmethod
     def from_dict(
         cls,
-        extremely_long_name_that_will_cause_wrapping_aaaaaaaaaaaaaaaaa: typing.Optional[
+        {_LONG_NAME}: typing.Optional[
             int
         ] = None,
     ) -> "TModel":
@@ -364,8 +469,8 @@ class TModel({_EXPECTED_MODEL_BASE}):
         Construct from a dictionary (eg. a POST payload).
 
         Args:
-            extremely_long_name_that_will_cause_wrapping_aaaaaaaaaaaaaaaaa: The
-                extremely_long_name_that_will_cause_wrapping_aaaaaaaaaaaaaaaaa
+            {_LONG_NAME}: The
+                {_LONG_NAME}
                 of the Model.
 
         Returns:
@@ -406,11 +511,11 @@ class TModel({_EXPECTED_MODEL_BASE}):
         ...
 
 
-Model: TModel = models.Model  # type: ignore
+Model: typing.Type[TModel] = models.Model  # type: ignore
 ''',
         ),
     ],
-    ids=["single", "multiple", "black"],
+    ids=["single", "single object", "multiple", "black"],
 )
 @pytest.mark.models_file
 def test_integration(schemas, expected_source):
@@ -475,7 +580,7 @@ class TModel({_EXPECTED_MODEL_BASE}):
     query: orm.Query
 
     # Model properties
-    id: typing.Optional[int]
+    id: "sqlalchemy.Column[typing.Optional[int]]"
 
     def __init__(self, id: typing.Optional[int] = None) -> None:
         """
@@ -533,7 +638,7 @@ class TModel({_EXPECTED_MODEL_BASE}):
         ...
 
 
-Model: TModel = models.Model  # type: ignore
+Model: typing.Type[TModel] = models.Model  # type: ignore
 '''
     assert source == expected_source
 
@@ -589,8 +694,8 @@ class TModel({_EXPECTED_MODEL_BASE}):
     query: orm.Query
 
     # Model properties
-    name: typing.Optional[str]
-    id: typing.Optional[int]
+    name: "sqlalchemy.Column[typing.Optional[str]]"
+    id: "sqlalchemy.Column[typing.Optional[int]]"
 
     def __init__(
         self, name: typing.Optional[str] = None, id: typing.Optional[int] = None
@@ -654,7 +759,7 @@ class TModel({_EXPECTED_MODEL_BASE}):
         ...
 
 
-Model: TModel = models.Model  # type: ignore
+Model: typing.Type[TModel] = models.Model  # type: ignore
 '''
     assert source == expected_source
 
@@ -706,7 +811,7 @@ class TModel({_EXPECTED_MODEL_BASE}):
     query: orm.Query
 
     # Model properties
-    id: typing.Optional[int]
+    id: "sqlalchemy.Column[typing.Optional[int]]"
 
     def __init__(self, id: typing.Optional[int] = None) -> None:
         """
@@ -764,6 +869,92 @@ class TModel({_EXPECTED_MODEL_BASE}):
         ...
 
 
-Model: TModel = models.Model  # type: ignore
+Model: typing.Type[TModel] = models.Model  # type: ignore
 '''
     assert source == expected_source
+
+
+def _generate_source(schemas, names):
+    """Generate the models file source from a schema."""
+    models = models_file.ModelsFile()
+    for schema, name in zip(schemas, names):
+        models.add_model(schema=schema, name=name)
+    return models.generate_models()
+
+
+def _create_source_file(source, tmp_path):
+    """Create a file with the source code."""
+    directory = tmp_path / "models"
+    directory.mkdir()
+    source_file = directory / "models.py"
+    source_file.write_text(source)
+    return source_file
+
+
+@pytest.mark.parametrize(
+    "schemas, names",
+    [
+        pytest.param(
+            [{"properties": {"id": {"type": "integer"}}}], ["Model"], id="simple",
+        ),
+        pytest.param(
+            [
+                {"properties": {"id": {"type": "integer"}}},
+                {"properties": {"model": {"type": "object", "x-de-$ref": "RefModel"}}},
+            ],
+            ["RefModel", "Model"],
+            id="object",
+        ),
+    ],
+)
+@pytest.mark.models_file
+def test_generate_type_return(tmp_path, schemas, names):
+    """
+    GIVEN schema
+    WHEN the models file is generated and mypy is run over it
+    THEN no errors are returned.
+    """
+    source = _generate_source(schemas, names)
+    source_file = _create_source_file(source, tmp_path)
+
+    _, _, returncode = api.run([str(source_file)])
+
+    assert returncode == 0
+
+
+@pytest.mark.parametrize(
+    "schema, mypy_check, expected_out_substr",
+    [
+        pytest.param(
+            {"properties": {"id": {"type": "integer"}}},
+            "reveal_type(Model.id)",
+            "'sqlalchemy.sql.schema.Column[Union[builtins.int, None]]'",
+            id="nullable column",
+        ),
+        pytest.param(
+            {"properties": {"id": {"type": "integer"}}},
+            "model = Model()\nreveal_type(model.id)",
+            "'Union[builtins.int, None]'",
+            id="nullable column instance",
+        ),
+        pytest.param(
+            {"properties": {"id": {"type": "integer", "nullable": False}}},
+            "reveal_type(Model.id)",
+            "'sqlalchemy.sql.schema.Column[builtins.int*]'",
+            id="not nullable column",
+        ),
+    ],
+)
+@pytest.mark.models_file
+def test_generate_type_check(tmp_path, schema, mypy_check, expected_out_substr):
+    """
+    GIVEN schema, a mypy check and expected mypy output substring
+    WHEN the models file is generated and mypy is run over it
+    THEN the expected output substring is in the mypy output.
+    """
+    source = _generate_source([schema], ["Model"]) + f"\n{mypy_check}"
+    source_file = _create_source_file(source, tmp_path)
+
+    out, _, _ = api.run([str(source_file)])
+
+    assert expected_out_substr in out